transfered fslech from cvs to git
here is the cvs log message

Initial Savannah checkin of the Metafont source for Elmar Bartel's
"fselch" chess font, as used in xboard and WinBoard.  Previously this
was available only by emailing Elmar Bartel or Tim Mann and asking for
a copy.  I'm checking it in to Savannah to make it easier to come by.

	 --Tim Mann