Change accelerators again to be more mac like
After playing around with the beta quite a bit, it really bothered me
to have half be ‘cmd’ and half be ‘ctrl’, to the point where it almost
made me want to do away with ‘cmd’ altogether  (really non-standard on
OSX). This way is a little smarter and only changes the conflicting
accelerators to alt-cmd which while being a little more mac standard,
as well as be acceptable ascetically when combined with other
accelerators that are just cmd.